pagination-comparision/requirements.txt
Flask==3.1.3
pymongo==4.16.0
python-dotenv==1.2.2
faker==40.13.0
requests==2.33.1